pyproject.toml
[project]
name = "music-history-ontology"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "spotipy>=2.20.0",
    "python-dotenv>=1.0.0",
    "pre-commit (>=4.1.0,<5.0.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "musicbrainzngs (>=0.7.1,<0.8.0)",
    "ratelimit (>=2.2.1,<3.0.0)",
    "rdflib (>=7.1.4,<8.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "coverage-badge>=1.1.2"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
